refactor(cli): reorganize imports and improve file structure for better maintainability
