+++ tests/cli/infrastructure/AwsSsmSecretProvider.test.ts
@@ -4,7 +4,7 @@ import {
   type StartedLocalStackContainer,
 } from '@testcontainers/localstack';
 import { afterAll, beforeAll, describe, expect, it } from 'vitest';
-import { AwsSsmSecretProvider } from '../../../src/cli/infrastructure/AwsSsmSecretProvider';
+import { AwsSsmSecretProvider } from '../../../src/cli/infrastructure/Aws/AwsSsmSecretProvider';
 
 const LOCALSTACK_IMAGE = 'localstack/localstack:stable';
 const PARAM_NAME = '/test/secret';
