+++ tests/cli/infrastructure/PackageJsonFinder.test.ts
@@ -1,7 +1,7 @@
 import { access, mkdir, rmdir, unlink, writeFile } from 'node:fs/promises';
 import { join } from 'node:path';
 import { afterAll, beforeEach, describe, expect, it } from 'vitest';
-import { PackageJsonFinder } from '../../../src/cli/infrastructure/PackageJsonFinder';
+import { PackageJsonFinder } from '../../../src/cli/infrastructure/VersionFinder/PackageJsonFinder';
 
 describe('PackageJsonFinder', () => {
   const tempDir = join(__dirname, 'temp-pkgjson-test');
